% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barsTriees.R
\name{barsTriees}
\alias{barsTriees}
\title{barsTriees
une fonction retournant  un barplot triee ggplot.}
\usage{
barsTriees(data, var)
}
\arguments{
\item{data}{un dataframe}

\item{var}{une variable qualitative}
}
\description{
barsTriees
une fonction retournant  un barplot triee ggplot.
}
