% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boitemoustache}
\alias{boitemoustache}
\title{Boxplot
Permet de comparer des boxplot classique a ceux de ggplot/plotly}
\usage{
boitemoustache(
  data,
  variable,
  nbClasses = 1,
  varwidth = FALSE,
  type = "classique",
  color = "grey",
  modalite = NULL
)
}
\arguments{
\item{data}{un dataframe.}

\item{variable}{une variable quanti du data.}

\item{nbClasses}{le nombre de classes pour cut la variable quanti (si =1 , pas de classes).}

\item{varwidth}{TRUE pour ajuster la taille des box selon la quantite de donnees.}

\item{type}{le type du graphique (ggplot,classique,ou plotly).}

\item{color}{la couleur des boxplot.}

\item{modalite}{variable qualitative a discriminer.}
}
\description{
Boxplot
Permet de comparer des boxplot classique a ceux de ggplot/plotly
}
