% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{densite}
\alias{densite}
\title{Density
Densite d'une variable quantitative selon le type de plot specifie}
\usage{
densite(data, variable, modalite = NULL, type = "classique", color = "black")
}
\arguments{
\item{data}{un dataframe}

\item{variable}{la variable}

\item{modalite}{variable a discriminer (Null par defaut)}

\item{type}{le type de graphique souhaite (classique,ggplot)}

\item{color}{la couleur du graphique}
}
\description{
Density
Densite d'une variable quantitative selon le type de plot specifie
}
