% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{diagbatons}
\alias{diagbatons}
\title{diagbatons
Permet de comparer un barplot classique a un ggplot/plotly.}
\usage{
diagbatons(
  data,
  modalite,
  tri = FALSE,
  color = "grey",
  type = "classique",
  variable = NULL
)
}
\arguments{
\item{data}{un dataframe.}

\item{modalite}{une variable qualitative.}

\item{tri}{si TRUE, trie les barres.}

\item{color}{la couleur.}

\item{type}{le type de graphique: ('classique','ggplot' ou 'plotly').}

\item{variable}{une variable quelconque.}
}
\value{
un barplot
}
\description{
diagbatons
Permet de comparer un barplot classique a un ggplot/plotly.
}
