% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramme.R
\name{histogramme}
\alias{histogramme}
\title{Histogramme d'une variable quantitative selon le type de plot specifie}
\usage{
histogramme(data, variable, nbClasses = 8, color = "grey", type = "classique")
}
\arguments{
\item{data}{un dataframe}

\item{variable}{la variable}

\item{nbClasses}{le nombre de classes souhaite pour decouper la variable}

\item{color}{la couleur du graphique}

\item{type}{le type de graphique souhaite ('classique','ggplot','plotly')}
}
\description{
Histogramme d'une variable quantitative selon le type de plot specifie
}
\examples{
histogramme(data = iris,variable = "Sepal.Length",nbClasses = 8,color = "red",type = "classique")
histogramme(data = iris,variable = "Sepal.Width",nbClasses = 12,color = "blue",type = "ggplot")
}
