% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic.R
\name{mosaique}
\alias{mosaique}
\title{Mosaicplot
Permet de comparer un mosaicplot classique a un ggplot/plotly.}
\usage{
mosaique(data, var1, var2, color = "grey", type = "classique")
}
\arguments{
\item{data}{un dataframe}

\item{var1}{une variable qualitative}

\item{var2}{une 2eme variable qualitative}

\item{color}{la couleur du graphique}

\item{type}{le type de graphique: ('classique','ggplot' ou 'plotly').}
}
\value{
un mosaicplot
}
\description{
Mosaicplot
Permet de comparer un mosaicplot classique a un ggplot/plotly.
}
