% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quali.R
\name{plot_quali}
\alias{plot_quali}
\title{permet de faire un plot qualitatif}
\usage{
plot_quali(
  plot_type = "barplot",
  data,
  var1,
  var2,
  tri = "non",
  color = "grey",
  type = "classique"
)
}
\arguments{
\item{plot_type}{le type de plot ('histogramme','Density','boxplot','scatterplot').}

\item{data}{un dataframe}

\item{var1}{une variable qualitative}

\item{var2}{une 2eme variable qualitative}

\item{tri}{tri un barplot si "oui".}

\item{color}{couleur du graphique.}

\item{type}{le type de graphe utilise ('classique','ggplot','plotly')}
}
\description{
permet de faire un plot qualitatif
}
