% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quanti_quali.R
\name{plot_quanti_quali}
\alias{plot_quanti_quali}
\title{Permet de faire des plots quanti vs quali}
\usage{
plot_quanti_quali(
  plot_type = "boxplot",
  data,
  var_quanti1,
  var_quanti2,
  var_quali,
  modalites = "toute la population",
  color = "grey",
  type = "classique"
)
}
\arguments{
\item{plot_type}{le type de plot ('boxplot','barplot','scatteplot').}

\item{data}{un dataframe.}

\item{var_quanti1}{une variable quantitative.}

\item{var_quanti2}{une variable quantitative (up: pas obligatoire).}

\item{var_quali}{une variable qualitative.}

\item{modalites}{choix de la modalite a discriminer.}

\item{color}{couleur du graphique.}

\item{type}{le type de graphique ("classique','ggplot','plotly')}
}
\description{
Permet de faire des plots quanti vs quali
}
