% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Scatterplot
Permet de comparer un scatterplot classique a un de ggplot}
\usage{
scatterplot(
  data,
  variable1,
  variable2,
  type = "classique",
  color = "black",
  pch = 19,
  modalite = NULL,
  reg = FALSE
)
}
\arguments{
\item{data}{un dataframe.}

\item{variable1}{une variable quanti.}

\item{variable2}{une variable quanti.}

\item{type}{le type du graphique (ggplot,classique ou plotly).}

\item{color}{la couleur des individus.}

\item{pch}{le caractère du point.}

\item{modalite}{la variable a dicriminer.}

\item{reg}{la courbe de regression lisse.}
}
\description{
Scatterplot
Permet de comparer un scatterplot classique a un de ggplot
}
